% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_helpers.R
\name{addInterfaceToPySODB}
\alias{addInterfaceToPySODB}
\title{addInterfaceToPySODB}
\usage{
addInterfaceToPySODB(env_to_use = "giotto_env")
}
\arguments{
\item{env_to_use}{conda environment into which PySODB should be installed.}
}
\description{
addInterfaceToPySODB

addInterfaceToPySODB
}
\details{
Installs PySODB into `env_to_use`.
*Note that manual installation is more stable; see \dontrun{getSODBDataset}.*
}
