% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_helpers.R
\name{check_py_for_pysodb}
\alias{check_py_for_pysodb}
\title{check_py_for_pysodb}
\usage{
check_py_for_pysodb(env_to_use = NULL)
}
\arguments{
\item{env_to_use}{conda environment in which the installation of PySODB should be checked.}
}
\description{
check_py_for_pysodb
}
\details{
Checks python environment for pysodb.
Prompts user to install if not found.
*Note that manual installation is more stable; see \dontrun{getSODBDataset}.*
}
