% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gDataDir}
\alias{gDataDir}
\title{Get GiottoData library path}
\usage{
gDataDir()
}
\description{
Utility function to find library install location of the GiottoData
package. Provided as shorthand to make loading files easier.
}
