% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_objects.R
\name{generate_mini_subobjects}
\alias{generate_mini_subobjects}
\title{Generate mini subobjects}
\usage{
generate_mini_subobjects()
}
\value{
NULL invisibly
}
\description{
Generate the GiottoData mini subobjects based on Mini_objects_script.R
in `/inst/Mini_objects/` subdirectory.
}
