% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{giottodata_paths}
\alias{giottodata_paths}
\alias{gdata_libdir}
\alias{gdata_subobject_libdir}
\alias{gdata_dataset_libdir}
\alias{gdata_devdir}
\alias{gdata_subobject_devdir}
\alias{gdata_dataset_devdir}
\alias{gdata_extdata_devdir}
\title{Get GiottoData paths}
\usage{
gdata_libdir(...)

gdata_subobject_libdir(...)

gdata_dataset_libdir(...)

gdata_devdir(...)

gdata_subobject_devdir(...)

gdata_dataset_devdir(...)

gdata_extdata_devdir(...)
}
\arguments{
\item{\dots}{passed to `file.path()`}
}
\description{
Utility functions to get helpful filepaths within the
\pkg{GiottoData} package.
}
\section{Functions}{
\itemize{
\item \code{gdata_libdir()}: Get the library install path
of the package. Should not be used in contexts where package is loaded with
`devtools::load_all()`

\item \code{gdata_subobject_libdir()}: Get the library path to the mini
subobjects directory

\item \code{gdata_dataset_libdir()}: Get the library path to the mini
datasets directory

\item \code{gdata_devdir()}: Get the development root path

\item \code{gdata_subobject_devdir()}: Get the development path to the mini
subobjects directory

\item \code{gdata_dataset_devdir()}: Get the development path to the mini
datasets directory

\item \code{gdata_extdata_devdir()}: Get the development path to the extdata
directory

}}
\keyword{internal}
