% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_objects.R
\name{list_subobject_mini}
\alias{list_subobject_mini}
\title{List available subobjects data}
\usage{
list_subobject_mini()
}
\description{
Utility function to list the available saved subobjects. This is
the internal function which returns a version of the table with the full pathfiles
and is not prettied or subsetted by type. See \code{\link{listSubObjectMini}}
for the external version which provides those functions.
}
\keyword{internal}
