% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_helpers.R
\name{listSODBDatasetNames}
\alias{listSODBDatasetNames}
\title{listSODBDatasetNames}
\usage{
listSODBDatasetNames(
  category = c("All", "Spatial Transcriptomics", "Spatial Proteomics",
    "Spatial Metabolomics", "Spatial Genomics", "Spatial MultiOmics")
)
}
\arguments{
\item{cateogry}{name of category for which dataset names will be listed.}
}
\description{
listSODBDatasetNames
}
\details{
Returns a vector containing the names of datasets associated with
the provided `category`.
}
