% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_objects.R
\name{listSubObjectMini}
\alias{listSubObjectMini}
\title{List available mini giotto subobjects}
\usage{
listSubObjectMini(x = NULL)
}
\arguments{
\item{x}{subobject type (NULL lists all subobject types)}
}
\description{
Lists the available mini giotto subobjects to load. Items are first
classified by the type of subobject, then by the index number which is used
to tell apart files when there is more than one example subobject of that type
available. The final column is the file name.
}
