% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_helpers.R
\name{getSODBDataset}
\alias{getSODBDataset}
\title{getSODBDataset}
\usage{
getSODBDataset(dataset_name = NULL, experiment_name = "default")
}
\arguments{
\item{dataset_name}{name of dataset to pull from the SODB. 
Must exist within the SODB.}

\item{experiment_name}{name of one experiment associated with `dataset_name`
By default, the first experiment will be used.}
}
\description{
getSODBDataset
}
\details{
Interface with TenCent's Spatial Omics DataBase (SODB) using the
python extension, pysodb.

This function will write an anndata h5ad file for a provided dataset
name to the current working directory and will then  convert
the h5ad into a Giotto Object.

Run \dontrun{listSODBDatasetNames()} to find names of SODB datasets.
Run \dontrun{listSODBDatasetExperimentNames()} to find names of
experiments associate with a provided dataset.
}
\examples{
\dontrun{

sodb_dataset_names = listSODBDatasetNames()
desired_dataset = sodb_dataset_names[[15]] # Arbitrary

dataset_experiment_names = listSODBDatasetExperimentNames(dataset_name = desired_dataset)
desired_experiment = dataset_experiment_names[[1]] # Arbitrary

gobject = getSODBDataset(dataset_name = desired_dataset,
                         experiment_name = desired_experiment)}
}
