% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_helpers.R
\name{getSpatialDataset}
\alias{getSpatialDataset}
\title{getSpatialDataset}
\usage{
getSpatialDataset(
  dataset = c("ST_OB1", "ST_OB2", "codex_spleen", "cycif_PDAC", "starmap_3D_cortex",
    "osmfish_SS_cortex", "merfish_preoptic", "mini_seqFISH", "seqfish_SS_cortex",
    "seqfish_OB", "slideseq_cerebellum", "ST_SCC", "scRNA_prostate", "scRNA_mouse_brain",
    "mol_cart_lung_873_C1", "sg_mini_kidney"),
  directory = getwd(),
  verbose = TRUE,
  dryrun = FALSE,
  ...
)
}
\arguments{
\item{dataset}{dataset to download}

\item{directory}{directory to save the data to}

\item{verbose}{verbosity}

\item{dryrun}{dryrun: does not download data but shows download commands}

\item{\dots}{additional parameters to \code{\link[utils]{download.file}}}
}
\description{
This function will automatically download the spatial locations and
expression matrix for the chosen dataset. These files are already in the right format
to create a Giotto object. If wget is installed on your machine, you can add
'method = wget' to the parameters to download files faster.
}
