% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_helpers.R
\name{listSODBDatasetExperimentNames}
\alias{listSODBDatasetExperimentNames}
\title{List SODB dataset experiment names}
\usage{
listSODBDatasetExperimentNames(dataset_name = NULL, env_name = NULL)
}
\arguments{
\item{dataset_name}{name of dataset for which experiment names will be listed.
Must exist within the SODB.}

\item{env_name}{Calls [GiottoClass::set_giotto_python_path()]. Path to
python binary to use or envname. Leaving as `NULL` will use the default.}
}
\description{
List SODB dataset experiment names
}
\details{
Returns a vector containing the names of experiments associated with
the provided `dataset_name`.

Run \preformatted{listSODBDatasetNames()} to find names of SODB datasets.
}
