% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_helpers.R
\name{listSODBDatasetExperimentNames}
\alias{listSODBDatasetExperimentNames}
\title{listSODBDatasetExperimentNames}
\usage{
listSODBDatasetExperimentNames(dataset_name = NULL)
}
\arguments{
\item{dataset_name}{name of dataset for which experiment names will be listed. 
Must exist within the SODB.}
}
\description{
listSODBDatasetExperimentNames
}
\details{
Returns a vector containing the names of experiments associated with
the provided `dataset_name`. 

Run \dontrun{listSODBDatasetNames()} to find names of SODB datasets.
}
