% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_helpers.R
\name{listSODBDatasetExperimentNames}
\alias{listSODBDatasetExperimentNames}
\title{listSODBDatasetExperimentNames}
\usage{
listSODBDatasetExperimentNames(dataset_name = NULL, env_name = "giotto_env")
}
\arguments{
\item{dataset_name}{name of dataset for which experiment names will be listed.
Must exist within the SODB.}

\item{env_name}{Python environment within which pysodb is installed.
If it is not already installed, the user
will be prompted to install `pysodb`
DEFAULT: "giotto_env"}
}
\description{
listSODBDatasetExperimentNames
}
\details{
Returns a vector containing the names of experiments associated with
the provided `dataset_name`.

Run \preformatted{listSODBDatasetNames()} to find names of SODB datasets.
}
