% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_helpers.R
\name{loadGiottoMini}
\alias{loadGiottoMini}
\title{loadGiottoMini}
\usage{
loadGiottoMini(
  dataset = c("visium", "seqfish", "starmap", "vizgen", "cosmx", "spatialgenomics"),
  python_path = NULL,
  init_gobject = TRUE,
  ...
)
}
\arguments{
\item{dataset}{mini dataset giotto object to load}

\item{python_path}{pythan path to use}

\item{init_gobject}{logical. Whether to initialize gobject on load}

\item{\dots}{additional params to pass to `GiottoClass::loadGiotto()`}
}
\description{
This function will automatically load one of the existing mini
giotto objects. These are processed giotto objects that can be used to test
Giotto functions and run examples. If no python path is provided it will try
to find and use the Giotto python environment.
Images associated with the giotto mini objects will be reconnected if possible.
Available datasets are:
\itemize{
  \item{1. visium: mini dataset created from the mouse brain sample }
  \item{2. vizgen: mini dataset created from the mouse brain sample }
  \item{3. cosmx: mini dataset created from the lung12 sample }
  \item{4. spatialgenomics: mini dataset created from the mouse kidney sample}
}
Instructions, such as for saving plots, can be changed
using the \code{\link{changeGiottoInstructions}}
}
