% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_subobjects.R
\name{loadSubObjectMini}
\alias{loadSubObjectMini}
\title{Load a mini giotto subobject}
\usage{
loadSubObjectMini(x, idx = 1L)
}
\arguments{
\item{x}{subobject type to load}

\item{idx}{which of multiple example subobjects to load when more than one
is available (see \code{\link{listSubObjectMini}})}
}
\description{
Utility function to load a mini giotto suite S4 subobject.
Can be useful for testing functions quickly or providing contained examples.
}
